Pull resources into their own component
